requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.5
pandas==2.2.3
pip==25.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rich==14.0.0
setuptools==75.8.0
six==1.17.0
tzdata==2025.2
wheel==0.45.1
